requirements.txt
streamlit==1.37.1
requests==2.32.3
Pillow==10.4.0
ImageHash==4.3.1
numpy==1.26.4
torch==2.2.1
torchvision==0.17.1
transformers==4.43.4